refactor: generic files moved into framework and especific files moved into application.
