+++ drivers/Web/framework/HttpRequest/Resource.py
