+++ drivers/Web/framework/router.py
@@ -1,8 +1,8 @@
 import logging
 import maybe
-from drivers.Web.http_response import HttpResponse
-from drivers.Web.HttpRequest.httpRequest import HttpRequest
-from drivers.Web.route_interface import Route
+from drivers.Web.framework.http_response import HttpResponse
+from drivers.Web.framework.HttpRequest.httpRequest import HttpRequest
+from drivers.Web.framework.route_interface import Route
 
 logger = logging.getLogger("drivers.Web.router")
